app/requirements.txt
Django==1.9.2
psycopg2==2.5.4
django-extensions==1.6.1
django-test-without-migrations==0.4
django-nose==1.4.2
coverage==4.0a1
boto==2.34.0
botocore==0.76.0
djangorestframework==3.3.2
gunicorn==19.4.5
shortuuid==0.4.2
django-ses==0.7.0
django-storages==1.1.8
twilio==5.4.0
phonenumbers==7.4.0
